pyproject.toml
[tool.poetry]
name = "uber-server"
version = "0.1.0"
description = ""
authors = ["Tomas Hanzlik <tomas.hanzlik0@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
fastapi-cache2 = "0.1.8"
uvicorn = "^0.18.2"
httpx = "^0.23.0"
requests = "^2.28.1"
loguru = "^0.6.0"
asgi-correlation-id = "^3.0.1"
slowapi = "^0.1.5"
respx = "^0.19.2"
gunicorn = "^20.1.0"
aioredis = "^2.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
freezegun = "^1.2.1"
locust = "^2.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
